Add base class for fibbage engines to reduce duplicate code. Add instance numbers to AI player names so that they can be identified in logging output. Update logging to allow prefixes, and use that to insert newlines between fibbage questions
