+++ src/Games/Common/Models/ISelectionChoice.cs
@@ -1,0 +1,7 @@
+﻿namespace JackboxGPT3.Games.Common.Models
+{
+    public interface ISelectionChoice
+    {
+        public string SelectionText { get; }
+    }
+}
