+++ src/Games/Fibbage3/Models/LieChoice.cs
@@ -1,13 +1,15 @@
-﻿using Newtonsoft.Json;
+﻿using JackboxGPT3.Games.Common.Models;
+using Newtonsoft.Json;
 
 namespace JackboxGPT3.Games.Fibbage3.Models
 {
-    public struct LieChoice
+    public struct LieChoice : ISelectionChoice
     {
         [JsonProperty("disabled")]
         public bool Disabled { get; set; }
 
         [JsonProperty("text")]
         public string Text { get; set; }
+        public string SelectionText => Text;
     }
 }
