+++ src/Games/Fibbage3/Models/TruthChoice.cs
@@ -1,8 +1,9 @@
-﻿using Newtonsoft.Json;
+﻿using JackboxGPT3.Games.Common.Models;
+using Newtonsoft.Json;
 
 namespace JackboxGPT3.Games.Fibbage3.Models
 {
-    public struct TruthChoice
+    public struct TruthChoice : ISelectionChoice
     {
         [JsonProperty("censorable")]
         public bool Censorable { get; set; }
@@ -18,5 +19,6 @@ namespace JackboxGPT3.Games.Fibbage3.Models
 
         [JsonProperty("text")]
         public string Text { get; set; }
+        public string SelectionText => Text;
     }
 }
